Using cec-client in tmux instead of python-cec
There where a lot of problems with using python-cec on a raspberry pi 4.
To solve this, libcec was compiled locally and cec-client was used instead.
There is a workarounda as the receiver won't listen to events until the unit has been an active device,
hence the as, standby 5, on 5 loop.